Add files via upload
Alien Attributes: The Alien class contains the following attributes:

x_coordinate and y_coordinate: These represent the alien's coordinates in a 2D space.
health: An integer value that represents the alien's health points, initially set to 3.
Alien Methods:

getHealth(): A method that retrieves the current health of the alien.
hit(): A method that simulates the alien taking damage by decrementing its health. It ensures that health does not go below zero.
is_alive(): A method that checks if the alien is still alive based on its health.
Teleportation: The program allows the alien to teleport to new coordinates by invoking the teleport(int x_new, int y_new) method.

Collision Detection: The collision_detection(const Alien& other_alien) method enables the alien to detect if it occupies the same coordinates as another alien object.

Main Function: The main function demonstrates the usage of the Alien class. It creates instances of Alien, simulates damage to the alien, checks its health status, teleports the alien, and checks for collisions with other aliens.